pyproject.toml
[project]
name = "deep_research"
version = "0.0.1"
description = "Answering a question with deep research agent."
authors = [
    { name = "Klaus Lochmann" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.5.4",
    "langchain-community>=0.3.9",
    "langchain-openai>=0.3.28",
    "langchain-anthropic>=0.3.15",
    "langchain-mcp-adapters>=0.1.6",
    "langchain-deepseek>=0.1.2",
    "langchain-tavily",
    "langchain-groq>=0.2.4",
    "openai>=1.99.2",
    "tavily-python>=0.5.0",
    "langgraph-cli[inmem]>=0.3.1",
    "langsmith>=0.3.37",
    "langchain-google-vertexai>=2.0.25",
    "langchain-google-genai>=2.1.5",
    "langchain>=1.0.3"
]